+++ src/main/java/com/library/rest/service/impl/BookServiceImpl.java
@@ -1,7 +1,9 @@
 package com.library.rest.service.impl;
 
+import com.library.rest.entities.Author;
 import com.library.rest.entities.Book;
 import com.library.rest.repository.BookRepository;
+import com.library.rest.service.AuthorService;
 import com.library.rest.service.BookService;
 import org.springframework.stereotype.Service;
 
@@ -11,9 +13,11 @@ import java.util.Optional;
 @Service
 public class BookServiceImpl implements BookService {
     private final BookRepository bookRepository;
+    private final AuthorService authorService;
 
-    public BookServiceImpl(BookRepository bookRepository) {
+    public BookServiceImpl(BookRepository bookRepository, AuthorService authorService) {
         this.bookRepository = bookRepository;
+        this.authorService = authorService;
     }
 
     @Override
@@ -28,16 +32,37 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public Book createBook(Book book) {
+        Long authorId  = book.getAuthor().getId();
+        Author author = authorService.getAuthorById(authorId).orElseThrow(() -> new RuntimeException("Cannot create book : Author not found with id : " + authorId));
+        book.setAuthor(author);
         return bookRepository.save(book);
     }
 
     @Override
     public Book updateBook(Long id, Book book) {
-        return null;
+        return bookRepository.findById(id)
+                .map(existingBook -> {
+                    existingBook.setTitle(book.getTitle());
+                    existingBook.setIsbn(book.getIsbn());
+                    existingBook.setPublishedYear(book.getPublishedYear());
+
+                    if (book.getAuthor() != null && book.getAuthor().getId() != null){
+                        Long authorId = book.getAuthor().getId();
+                        Author author = authorService.getAuthorById(authorId)
+                                .orElseThrow(() -> new RuntimeException("Cannot update book : No author found with id : " + authorId));
+                        existingBook.setAuthor(author);
+                    }
+                    return bookRepository.save(existingBook);
+                })
+                .orElseThrow(() -> new RuntimeException("Book not found with id : " + id ));
     }
 
     @Override
     public void deleteBook(Long id) {
-        bookRepository.deleteById(id);
+        if (bookRepository.existsById(id)){
+            bookRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Cannot delete : Book not found with id : " + id);
+        }
     }
 }
